Some General Tips About Solana Development
In Solana development, you’ll face a lot of weird custom errors and bugs, and it can be quite hard and frustrating to fix them since the Solana dev ecosystem is not as big as the Eth dev ecosystem. But not to worry. When you get stuck you simply have to look in the right place for the solution.
During my development, I was constantly asking my doubts in the Anchor discord server, the Metaplex and Superteam server, and looking through other code repos on GitHub and the Metaplex Program library itself.